Move `EditionDependenciesPopulator`
This is a `ServiceListener` so has been moved for consistency with all
of the others.
